Little more fixing up of the pack controller, and a bugfix in board repository.

+++ Controllers/PacksController.cs
@@ -76,30 +76,41 @@ namespace Howler.Controllers
         public IActionResult Update(int id, Pack pack)
         {
             User sender = GetCurrentUser();
+            User updatedPackLeader = _userRepository.GetById(pack.PackLeaderId); 
             Pack packBeingUpdated = _packRepository.GetById(id);
             Pack packWithSameName = _packRepository.ExactSearch(pack.Name);
-            if(packBeingUpdated.PackLeaderId != sender.Id) 
+            if(packBeingUpdated != null)
             {
-                return Forbid();
-            }
-            if(packBeingUpdated.PrimaryBoardId != pack.PrimaryBoardId)
-            {
-                return BadRequest();
-            }
-            if (sender.PackId != packBeingUpdated.Id || sender.PackId != pack.Id || pack.Id != packBeingUpdated.Id)
-            {
-                return BadRequest();
-            }
-            if(packWithSameName != null)
-            {
-                ObjectResult response = new ObjectResult(new { title = "Already Exists", status = 420, message = $"A pack named '{pack.Name}' already exists in the database" });
-                response.StatusCode = 420;
-                return response;
-            }
 
-            _packRepository.Edit(pack);
-            return NoContent();
+                if(packBeingUpdated.PackLeaderId != sender.Id) 
+                {
+                    return Forbid();
+                }
+                if(packBeingUpdated.PrimaryBoardId != pack.PrimaryBoardId)
+                {
+                    return BadRequest();
+                }
+                if (sender.PackId != packBeingUpdated.Id || sender.PackId != pack.Id || pack.Id != packBeingUpdated.Id || updatedPackLeader == null)
+                {
+                    return BadRequest();
+                }
+                if(updatedPackLeader.PackId != pack.Id || updatedPackLeader.PackId != packBeingUpdated.Id)
+                {
+                    ObjectResult response = new ObjectResult(new { title = "Can't Assign Pack to Non-Member", status = 470, message = $"{updatedPackLeader.DisplayName} is not a member of this pack, and cannot be set as the leader!" });
+                    response.StatusCode = 470;
+                    return response;
+                }
+                if(packWithSameName != null && packWithSameName.Id != pack.Id)
+                {
+                    ObjectResult response = new ObjectResult(new { title = "Already Exists", status = 420, message = $"A pack named '{pack.Name}' already exists in the database" });
+                    response.StatusCode = 420;
+                    return response;
+                }
 
+                _packRepository.Edit(pack);
+                return NoContent();
+            }
+            return NotFound();
         }
 
         [HttpDelete("{id}")]
@@ -107,16 +118,20 @@ namespace Howler.Controllers
         {
             User currentUser = GetCurrentUser();
             Pack packToDelete = _packRepository.GetById(id);
-            if(packToDelete.PackLeaderId != currentUser.Id)
+            if(packToDelete != null)
             {
-                return Forbid();
+                if(packToDelete.PackLeaderId != currentUser.Id)
+                {
+                    return Forbid();
+                }
+                if (packToDelete.PrimaryBoardId != null)
+                {
+                    _boardRepository.Delete((int)packToDelete.PrimaryBoardId);
+                }
+                _packRepository.Delete(id);
+                return NoContent();
             }
-            if (packToDelete.PrimaryBoardId != null)
-            {
-                _boardRepository.Delete((int)packToDelete.PrimaryBoardId);
-            }
-            _packRepository.Delete(id);
-            return NoContent();
+            return NotFound();
         }
 
         [HttpGet]
